+++ crates/biome_js_semantic/src/semantic_model/closure.rs
@@ -130,9 +130,9 @@ impl Iterator for AllCapturesIter {
     fn next(&mut self) -> Option<Self::Item> {
         'references: loop {
             while let Some(reference) = self.references.pop() {
-                let binding = &self.data.bindings[reference.binding_id];
+                let binding = &self.data.bindings[reference.binding_id as usize];
                 if self.closure_range.intersect(binding.range).is_none() {
-                    let reference = &binding.references[reference.reference_id];
+                    let reference = &binding.references[reference.reference_id as usize];
                     return Some(Capture {
                         data: self.data.clone(),
                         node: self.data.binding_nodes[&reference.range.start()].clone(), // TODO change node to store the range
